feat(server,worker): seed-anchored viz subset, float32 caches, and an Atlas byte cap for the crawler

+++ src/music_recommendations/server/viz.py
@@ -35,7 +35,7 @@ def clear_geometry_cache() -> None:
 
 
 def normalized_rows(matrix: np.ndarray) -> np.ndarray:
-    matrix = np.asarray(matrix, dtype=float)
+    matrix = np.asarray(matrix, dtype=np.float32)
     norms = np.linalg.norm(matrix, axis=1, keepdims=True)
     return matrix / np.where(norms == 0.0, 1.0, norms)
